Successfully modularized code, just need to refactor to get it working again
