+++ client/containers/button.jsx
@@ -1,0 +1,2 @@
+import React, { Component } from 'react';
+import '../styles.scss';
